backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
# Password hashing dependencies - pin to known-compatible versions for Python 3.13
# bcrypt 4.1.0 is known to provide the necessary backend metadata for passlib
# passlib[bcrypt] ensures the bcrypt backend integration
bcrypt==4.1.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator
# AI and Ollama dependencies
requests>=2.31.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0 